[Dev] Fix footgun in `string_t::SetSizeAndFinalize` (https://redirect.github.com/duckdb/duckdb/pull/18885)
The method was initially created only to truncate inlined strings, that
were already created as inlined.
It's now also used by codepaths that use `StringHeap::EmptyString` to
create a (potentially not inlined) string.

This size given to `EmptyString` could be bigger than the final size, if
the caller is attempting to reduce reallocations by allocating for the
potentially biggest result from the start.

In the event that the allocated size is `> INLINE_LENGTH` but the final
size is `<= INLINE_LENGTH`, this causes a problem because the data
written to the `value.pointer.ptr` is never copied to the
`value.inlined.inlined`.

I've added an extra parameter to `SetSizeAndFinalize`: `idx_t
allocated_size`, this should remove this footgun.

I've also added a test to the `test_bignum_sum.test` that fails on
`main`, because the result consisted of uninitialized inlined bytes.

Upstream-commit: https://github.com/duckdb/duckdb/commit/fd36bea9393cefafc323996f4adde370fb51e922