Encryption now encoded as a bit, centralizing in set/getter (https://redirect.github.com/duckdb/duckdb/pull/18897)
Follow up from https://github.com/duckdb/duckdb/pull/18823, in
particular comment at
https://github.com/duckdb/duckdb/pull/18823#discussion_r2317434418,
where we changed way to decide if a Database file is encrypted, but only
in a single place.

This PR propagate the change to the `IsEncrypted` function, so same
logic applies also to other codepath, and centralize also Setter to a
location close-by in the code.

This PR should be compatible / orthogonal to the other in-flight
encryption PR, but maybe that's best checked first merging the other
one.

Upstream-commit: https://github.com/duckdb/duckdb/commit/2718d1db52a3ecf8aed9858d9b7ce9b09b22d114