Http_util can return success for all [200, 300) responses, as well as redirects (https://redirect.github.com/duckdb/duckdb/pull/18940)
Currently only OK_200, and NotModified_304 return HTTPResponse objects
indicating success. With DuckDB iceberg, we have been using aws and
httputil to make requests, and the response status values have not been
consistent. I think all 200 responses should merit success, as well as
NotModified_304.

One example where this has come up is that Iceberg makes Head requests
to verify the existence of a schema. Many IRCs respond with
204_NoContent. The Http_util library will label this as success = false,
when that is indeed no the case.

Also happy to open this to v1.4-andium.

Upstream-commit: https://github.com/duckdb/duckdb/commit/81141ad556026089a2d29f3c7d7b537caa306934