Fix error reporting in SSLClient (https://redirect.github.com/duckdb/duckdb/pull/18958)
With `httplib`, when the `SSLClient` is used to connect to plain-HTTP
server (that can happen due to misconfiguration in `CREATE SECRET`), it
can return a failure from the `send()` call without setting the `Error`
reference to the corresponding error code. This can cause problems to
callers, that may expect that, when the check like this is passed on the
response:

```c++
if (res.error() == duckdb_httplib_openssl::Error::Success)
```

then they can access the response contents with `res.value()`. When
`SSLClient`'s connection fails - the contents `unique_ptr` is not set
and an attemt to access it causes runtime assertion:

```
Attempted to dereference unique_ptr that is NULL!
```

This change fixes the `SSLClient::create_and_connect_socket` method
making sure that, the `Error` value is set correctly when the
`is_valid()` check fails.

Fixes: duckdblabs/duckdb-internalhttps://redirect.github.com/duckdb/duckdb/pull/5865
Upstream-commit: https://github.com/duckdb/duckdb/commit/ab2ff61ad978bc9da668df55bad28d05c1bd1940