Allow extensions to customize ATTACH OR REPLACE conflict behavior (https://redirect.github.com/duckdb/duckdb/pull/18962)
The `ATTACH OR REPLACE` command currently is a NOOP in case there is
already an attached database with the same path and database type. It
disregards other attach options, however.

In case where path + database type is not uniquely characterizing a
database, this means that the reattach is not taking place.

This PR makes the behavior extensible, so that this behavior can be
extended to check for other attach options.

Upstream-commit: https://github.com/duckdb/duckdb/commit/551a5a537b8d6f88f28e32e7abf1a06edd2e90a8