Unify test runner keyword replacement, and don't run `LOAD [ext]` by default (https://redirect.github.com/duckdb/duckdb/pull/18963)
This PR fixes two issues with the test runner:

* We unify keyword replacement between SQL queries (supporting
`__TEST_DIR__` and friends) and the test config (supporting
`{BASE_TEST_NAME}` and friends). This should all just go through the
same path. Drawback here is that the syntax is different. I would
propose we should move to using `{TEST_DIR}` over `__TEST_DIR__` going
forward - but we can tackle that another day.
* Avoid running `LOAD 'ext'` for every extension with default
autoloading settings - this triggers exceptions at the beginning of
every test run which is kind of annoying if you are adding breakpoints
on exceptions.

Upstream-commit: https://github.com/duckdb/duckdb/commit/306c3caf86216d4a9a18ff144dd7951e8018a0db