Correctly update row group data pointers and root table pointer after checkpoint (https://redirect.github.com/duckdb/duckdb/pull/18966)
This PR fixes an issue with the experimental metadata re-use.

When running a `CHECKPOINT`, we would check if a table was modified or
not. If a table was unmodified, we would write the root pointer from
where we loaded the table instead of rewriting the table. When the table
was modified, we would re-write the table data to the metadata.

The issue here was that if we would first run a `CHECKPOINT` in which we
would rewrite the table, we would re-write the table data, but not
update the in-memory pointer to the new location for the table metadata.
If we would then do another `CHECKPOINT` in which the table was
unmodified - we would write a pointer to the (old) metadata instead of
writing a pointer to the new metadata. This could then upon restart
cause the table to point to old metadata, which could cause various
issues as that data could have been already overwritten and now contain
different data.

This PR addresses this issue by correctly overwriting the in-memory
pointers of the row groups and the table metadata, which resolves the
issue.

Upstream-commit: https://github.com/duckdb/duckdb/commit/a00499ff6c59fa2186a910a0f2a4fe9d78b26442