Attach: Cleanup duplicate data path handling, and make IF NOT EXISTS no longer abort if we are adding a path with the same name (https://redirect.github.com/duckdb/duckdb/pull/18974)
This PR performs some clean-up in duplicate data path handling by
centralizing the tracking of database paths in a single location, and
ensuring they are always cleaned up correctly through the use of RAII.
In addition, when we encounter a conflict we check if we are running
with `ATTACH IF NOT EXISTS` and are trying to attach the same name. If
so, instead of throwing an error, we succeed the attach.

Upstream-commit: https://github.com/duckdb/duckdb/commit/4659028e3a877f7134cee56d75b2535f053df795