Check context.interrupted flag in table scan (https://redirect.github.com/duckdb/duckdb/pull/18981)
There's a current design flaw in the TableFunction API: because an empty
chunk signals that we have no more output, we need to always return data
before handing back control upstream. This is usually not a problem, but
when we have very selective filters, this can cause us to spend a long
time in the table scan. We might be reading the entire table to find a
single row, never handing back control until we have found that row.

This PR introduces a work-around to make query interruption work in this
case by checking the interrupted flag. However, there's still other
problems with not handing back control (scheduling fairness, etc). We
need to address this separately but this requires an API change - so
this will be left for a future PR.

Upstream-commit: https://github.com/duckdb/duckdb/commit/b3edbac8519f8ed04f58a6f30ec349112bdc7d6c