ATTACH IF NOT EXISTS - wait until database is fully attached before returning (https://redirect.github.com/duckdb/duckdb/pull/18993)
https://github.com/duckdb/duckdb/pull/18974 cleaned up handling of
`ATTACH IF NOT EXISTS` - where instead of throwing if a database with
the same path/name was in the process of being attached, we would return
without erroring.

However, this behavior is likely still undesirable - since we would
return while the database was *in process of being attached by a
different connection* but not yet fully attached/queryable, we could get
situations like this:

```sql
ATTACH IF NOT EXISTS 'db_file.db' AS db_name;
CREATE TABLE db_name.tbl(i INTEGER);
-- Database with name "db_name" does not exist
```

This PR resolves this by instead busy spinning until either (1) the
database is actually attached, or (2) the database failed to attach in
the other connection, in which case we can continue and try our attach.

Upstream-commit: https://github.com/duckdb/duckdb/commit/7efeba3391a68809c06bc8eb226eeec960c746bb