WALReplay Fix: In UpdateColumn, no longer assume all updates are part of the same vector, but instead verify this and batch updates per vector (https://redirect.github.com/duckdb/duckdb/pull/18999)
Since https://github.com/duckdb/duckdb/pull/18829, it is easier for row
groups to be less strictly aligned, and also for the same series of
appends to not necessarily create the exact same row group layout.

The current implementation for update replay in the WAL requires row ids
to be perfectly stable: it assumes that it can figure out based on the
first row-id in the vector the vector that the rest of the updates
belong to. This PR reworks the `UpdateColumn` method to instead figure
out the correct vector that each of the row-ids in the update belongs to
prior to performing the update. As a result, this code is robust to
differences in row group layouts and the row-ids are no longer required
to all be part of the same vector.

Upstream-commit: https://github.com/duckdb/duckdb/commit/d232ee0ccb5abb2ce34a806c7408772957fc9d24