Fix threading issues in metadata manager, and expand concurrent attach / detach fuzz test (https://redirect.github.com/duckdb/duckdb/pull/19054)
This PR fixes a threading issue in the Metadata Manager caused by
performing reads while checkpointing. In particular, reads can trigger
metadata loads due to lazy metadata reading - which can cause a race
with the checkpointing code that writes new metadata to disk.

In addition, we expand the attach / detach test to do more operations
(such as checkpointing and delete + inserts in a single transaction) to
help detect more issues.

Upstream-commit: https://github.com/duckdb/duckdb/commit/4039239f1024468b2c6fa38143080ee1a5a9af4e