Always execute cast and try_cast if they are not invertible (https://redirect.github.com/duckdb/duckdb/pull/19010)
fixes https://github.com/duckdblabs/duckdb-internal/issues/5876 
fixes https://github.com/duckdb/duckdb/issues/18967

Together with @Tmonster we managed to come up with a fix. We identified
two separate issue.
- an inconsistency when running the same query with and without a
conjuction. The problem there was that even though we were making
changes in the child expression, the variable changes_made was never
updated. As a result, the expression rewriter was not aware of the
changes and was not re-reruning before going to the next rule.
- Even when the cast is not invertible, the code would drop the cast and
just compare the LHS and RHS if the RHS could be casted (and invertible)
to the target type.

To solve this we identified two solutions. One option is to change the
functionality and when the cast is not invertible always run the cast or
try_cast, but that is more work for the engine. The other option is to
keep the old functionality, and just fix the inconsistency

Upstream-commit: https://github.com/duckdb/duckdb/commit/eaa28d602f505d0bb6a2704bf6a379573455dccb