Skip serialization of row_start in DataPointer when targeting latest storage (https://redirect.github.com/duckdb/duckdb/pull/19111)
Writing the start row for column segments is not useful - since we
already have the start position of the row group, we can just set the
start row for each column segment to the correct location. This is now
done already as well in https://github.com/duckdb/duckdb/pull/19055 as
this actually resolved a bug with these start positions being written
incorrectly.

This PR removes the start row positions when targeting the latest
storage version. As a result, we write less metadata at no loss of
information. For example, for `hits`, the total metadata size decreases
from 14MB to 12MB as a result of this change.

Upstream-commit: https://github.com/duckdb/duckdb/commit/014e36ab0b60d1698b61d2a71314c6babdbaa574