[C-API] Initial support for accessing file system (https://redirect.github.com/duckdb/duckdb/pull/19086)
This PR adds initial support for accessing DuckDB's filesystem layer
from the C-API.

With the current state of the C-API this is primarily useful when
writing table functions that scan some external file format, as it
allows you to make use of DuckDBs existing cross platform filesystem, as
well as any loaded extension file systems (like httpfs), without having
to reimplement the equivalent functionality yourself.

I've tried to keep this initial version of the interface as small and
simple as possible, while also leaving room to expand it in the future.
The most complex part is the error handling, but I think I've solved it
relatively neatly by having both the `duckdb_file_system` and
`duckdb_file_handle` store their last occurred error internally, which
can then be converted to `duckdb_error_data` through a pair of separate
functions. Other than that the read/write functions follow posix
semantics, returning the number of bytes read/written, 0 on EOF and a
negative value on error. In practice the error value is always `-1`, but
I've left it unspecified so that we can potentially return specific
error codes in the future (like something like EAGAIN, when we get
asyncio), without forcing the user to allocate, inspect and destroy a
`duckdb_error_data` object to handle "common" IO errors.

Upstream-commit: https://github.com/duckdb/duckdb/commit/8d72b0ef6115a3262fa49c9570593782d7d17626