Fix unholy interaction between `CREATE VIEW` and `CTE` definitions (https://redirect.github.com/duckdb/duckdb/pull/19116)
This PR fixes an exponential growth that could occur under certain
circumstances when `VIEW` definitions used `CTE`s.

To fix this, the `cte_map` must no longer be copied in the transformer,
when a `CTE` node is created.
This required a few further changes to `bind_basetableref`, as we can no
longer access the `CommonTableExpressionInfo` entry. But, as it turns
out, this is no longer required anyways.

Upstream-commit: https://github.com/duckdb/duckdb/commit/b936ce8f545564708d368b28214d5a700c91ae3e

---

Fix: https://github.com/duckdb/duckdb/issues/19067