Fix internal issue 5975 (https://redirect.github.com/duckdb/duckdb/pull/19101)
Fixes https://github.com/duckdblabs/duckdb-internal/issues/5975

The offending lines are just under `if (!op.perform_delim) {`, which
assume that `op.correlated_columns[0]` is the `delim_index`, but the
test case that I've added has a different column there. I've fixed this
by looping through the columns to find the `delim_index`, and now the
test succeeds.

Not sure if this is the right way to do it @kryonix 

Upstream-commit: https://github.com/duckdb/duckdb/commit/8f570dd861f89d13f7c15322c0072af44509358b