Introduce `parser_override_function_t` (https://redirect.github.com/duckdb/duckdb/pull/19126)
This is a follow-up/improvement over
https://github.com/duckdb/duckdb/pull/18937

It adds a new function pointer `parser_override_function_t` as part of
the `ParserExtension` with which extensions can override the current
parser. This hook is executed *before* the current parser and takes as
input the entire query string, which can contain multiple statements and
returns a vector of (potentially multiple) statement(s). The
`ParserOverride` can also return either (1) a custom error message which
is turned into a `ParserException` (2) Failed parse in which case
DuckDB's parser can try to parse & transform the query.

In a follow-up PR I will add the PEG parser which will make use of this
new hook.

Upstream-commit: https://github.com/duckdb/duckdb/commit/4bfb6e2f8c74c7a02e25ca80bad68456270e545b