Delay throwing `NotImplementedException` in `ExpressionBinder` (https://redirect.github.com/duckdb/duckdb/pull/19153)
Fixes https://github.com/duckdb/duckdb/issues/19119

Since v1.4.0 we bind macro arguments to determine their type. We do this
using a temporary `ExpressionBinder` that we create, which is not able
to bind `ExpressionClass::WINDOW`, which only `SelectBinder` can do.

We already have a path to success for expressions that we cannot bind:
default to the previous (pre v1.4.0) behaviour by using
`LogicalType::UNKNOWN`. However, the `ExpressionBinder` immediately
would immediately throw, instead of wrapping the error in a
`BindResult`. This PR wraps the error instead of throwing immediately,
allowing us to detect this and continue as before.

Upstream-commit: https://github.com/duckdb/duckdb/commit/c9ec4e2ad550df43bf719cdfb8af6a68e89e059b