ATTACH IF NOT EXISTS: avoid looping waiting for DETACH to finish, wait only for an ATTACH operation to finish (https://redirect.github.com/duckdb/duckdb/pull/19193)
Waiting for an `ATTACH` to finish is not very problematic as this
generally involves only starting up the database. Detaches are more
problematic since a detach is only finished when all transactions that
refer to a given database are closed - hence this could end up taking a
very long time / forever.

Upstream-commit: https://github.com/duckdb/duckdb/commit/54de81b2861feee5a663320f28d598948f211fae