Disable ALP for non-default block sizes (https://redirect.github.com/duckdb/duckdb/pull/19197)
Currently, the ALP code assumes that it can always write 1024 values to
a block. That might not be the case in the presence of many exceptions,
if block sizes are small. Eventually, we should refactor the ALP
compression code to address that, see
https://github.com/duckdblabs/duckdb-internal/issues/6076. For now, we
just disable it for non-default block sizes.

Fix https://github.com/duckdblabs/duckdb-internal/issues/6075.

Upstream-commit: https://github.com/duckdb/duckdb/commit/3a557fe42ff35c53351daed9e30cd88e2f039efd