Cross and lateral join pushdown with unnest or json - update  (https://redirect.github.com/duckdb/duckdb/pull/19152)
After @Tmonster pointed it out in my previous PR, I moved the logic for
the ```unnest``` filter pushdown to a more suitable place to avoid
unnecessary checks in ```FinishPushdown``` which is called multiple
times.

In addition, I re-activated the test case that was failing only on linux
before, but removed the ```where``` clause. The problem was caused
because now the filter is pushed down below the ```delim join```, so the
```unnest``` rewriter optimization no longer takes place.

Upstream-commit: https://github.com/duckdb/duckdb/commit/dbc910cdfab25648838c6db2b3068b97ec9688c9