Common Subplan Elimination (https://redirect.github.com/duckdb/duckdb/pull/19080)
This PR implements the `CommonSubplanOptimizer`, which detects common
subplans and eliminates them by materializing them as a CTE, and
replacing the subplans with references to the CTE.

## Implementation

The optimizer uses bottom-up recursion to build up and identify
equivalent subplans.

### Detection

Detecting common subplans starts at leaf nodes such as `LogicalGet`,
converts them to a _"canonical representation"_, and then adds them to a
hash map.

No two operators are ever equal in DuckDB, mainly due to having
different _table indices_ (and sometimes other minor differences).
Therefore, operators are converted to a canonical representation that
equalizes these table indices, _i.e._, makes them start at over at 0, 1,
etc., making them compare equal.

We don't want to mangle query plans with these conversions, so we only
temporarily convert the operators to this representation, then serialize
them to a blob using the `BinarySerializer`, and then restore the
original operator. I've called this blob serialization of the canonical
representation of an operator a _"signature"_. If signatures are equal,
the operators are equivalent.

I initially started implemented custom comparison code for each
operators, but this was very cumbersome, and will be hard to maintain in
the future, and newly added fields will have to be added to the
comparisons. Our serialization is well-tested, so reusing it makes this
optimizer much more robust to future changes.

### Replacement

After detection, a `LogicalMaterializedCTE` is placed at the lowest
common ancestor of the common subplan, and the common subplan is placed
as the left child. Then, the common subplans are replaced with
`LogicalCTERef`s to the newly created materialized CTE.

Finally, we use the `ColumnBindingReplacer` to replace the
`ColumnBinding`s that used to refer to bindings from the common subplans
with bindings that refer to the columns coming from the CTE reference.

### Performance

I have spent some time improving the performance of this optimizer. I've
reduced the amount of allocations by writing all signatures to a single
`MemoryStream` allocation. Hashes of these signatures are computed once
and are cached.

With these performance improvements, the time it takes to detect common
subplans does not really stand out compared to other optimizer passes.
It takes just slightly longer than, e.g., the `FilterPushdown`
optimizer.

### Limitations

There are some limitations to the current implementation that can be
fixed in a future PR:
1. The optimizer will only eliminate one subplan: the largest subplan
(most operators) that occurs more than once.
2. The optimizer reuses the (rather basic) rules from the `CTEInlining`
optimizer to decide which common subplans should be eliminated.
3. If a CTE was marked as `NOT MATERIALIZED` through SQL, this optimizer
may detect the common subplan and decide to materialize it anyway.

## Example

Take, for example, the following query:
```sql
select t1.s + t2.s
from (select sum(range) s from range(10)) t1,
     (select sum(range) s from range(10)) t2
```
With corresponding query plan:
```
┌───────────────────────────┐
│         PROJECTION        │
│    ────────────────────   │
│          (s + s)          │
│                           │
│           ~1 row          │
└─────────────┬─────────────┘
┌─────────────┴─────────────┐
│       CROSS_PRODUCT       ├──────────────┐
└─────────────┬─────────────┘              │
┌─────────────┴─────────────┐┌─────────────┴─────────────┐
│    UNGROUPED_AGGREGATE    ││    UNGROUPED_AGGREGATE    │
│    ────────────────────   ││    ────────────────────   │
│    Aggregates: sum(https://redirect.github.com/duckdb/duckdb/pull/0)    ││    Aggregates: sum(https://redirect.github.com/duckdb/duckdb/pull/0)    │
└─────────────┬─────────────┘└─────────────┬─────────────┘
┌─────────────┴─────────────┐┌─────────────┴─────────────┐
│         PROJECTION        ││         PROJECTION        │
│    ────────────────────   ││    ────────────────────   │
│           range           ││           range           │
│                           ││                           │
│          ~10 rows         ││          ~10 rows         │
└─────────────┬─────────────┘└─────────────┬─────────────┘
┌─────────────┴─────────────┐┌─────────────┴─────────────┐
│           RANGE           ││           RANGE           │
│    ────────────────────   ││    ────────────────────   │
│      Function: RANGE      ││      Function: RANGE      │
│                           ││                           │
│          ~10 rows         ││          ~10 rows         │
└───────────────────────────┘└───────────────────────────┘
```
The query plan contains a common subplan corresponding to `select
sum(range) s from range(10)` in the query. With the new optimizer, this
query plan is converted to:
```
┌───────────────────────────┐
│         PROJECTION        │
│    ────────────────────   │
│          (s + s)          │
│                           │
│           ~1 row          │
└─────────────┬─────────────┘
┌─────────────┴─────────────┐
│            CTE            │
│    ────────────────────   │
│         CTE Name:         │
│     __common_subplan_1    │
│                           ├──────────────┐
│      Table Index: 20      │              │
│                           │              │
│           ~1 row          │              │
└─────────────┬─────────────┘              │
┌─────────────┴─────────────┐┌─────────────┴─────────────┐
│    UNGROUPED_AGGREGATE    ││       CROSS_PRODUCT       │
│    ────────────────────   ││                           ├──────────────┐
│    Aggregates: sum(https://redirect.github.com/duckdb/duckdb/pull/0)    ││                           │              │
└─────────────┬─────────────┘└─────────────┬─────────────┘              │
┌─────────────┴─────────────┐┌─────────────┴─────────────┐┌─────────────┴─────────────┐
│         PROJECTION        ││          CTE_SCAN         ││          CTE_SCAN         │
│    ────────────────────   ││    ────────────────────   ││    ────────────────────   │
│           range           ││       CTE Index: 20       ││       CTE Index: 20       │
│                           ││                           ││                           │
│          ~10 rows         ││           ~1 row          ││           ~1 row          │
└─────────────┬─────────────┘└───────────────────────────┘└───────────────────────────┘
┌─────────────┴─────────────┐
│           RANGE           │
│    ────────────────────   │
│      Function: RANGE      │
│                           │
│          ~10 rows         │
└───────────────────────────┘
```
This is a rather simple example, but the optimizer can detect much more
complex patterns, such as the ones in TPC-DS query 44 and 65.

Upstream-commit: https://github.com/duckdb/duckdb/commit/f635c7886e4221570aafc1fc909f488a53aea600