Inconsistent semicolon handling (https://redirect.github.com/duckdb/duckdb/pull/19199)
Fixes https://github.com/duckdblabs/duckdb-internal/issues/5474

The previous logic for splitting a multi-statement query (containing
multiple queries separated by `;`) was flawed. The split was
conditional, based on whether the preceding token was a
`SIMPLIFIED_TOKEN_OPERATOR` (like `)`). This logic led to the semicolon
being inconsistently included or excluded, resulting in unreliable
behavior for parser extensions.

This PR simplifies the query splitting process to be consistent. Instead
of the previous logic, we use the `StringUtil::Split()` to split the
full query string by the semicolon (`;`). Then we re-append the `;` to
every separate query.

Upstream-commit: https://github.com/duckdb/duckdb/commit/9244979e0ae9b220900037331d456a342aab6d56