[Transaction] Delete and drop of a table can now happen in the same transaction without error (https://redirect.github.com/duckdb/duckdb/pull/18918)
This PR fixes https://github.com/duckdblabs/duckdb-internal/issues/5632

### Problem description

Back in https://github.com/duckdb/duckdb/pull/17018 we introduced logic
to detect when a table had been dropped by a different transaction, so
we can properly error when deletes/updates/inserts are done on this
table.

But that introduced this regression, which caused deletes and drops to
no longer be allowed on the same table in the same transaction.

### Summary of changes
- `RowGroupCollection::Update` now forwards this `DataTable &` down to
the `RowGroup`, and the `ColumnData` and the `UpdateSegment`, to end up
in the `UpdateInfo`.
- `modified_tables` is no longer used in `DuckTransaction::Commit` to
check for updates/deletes/inserts made to deleted tables, instead this
check is moved to the commit of the `INSERT_TUPLE`, `UPDATE_TUPLE` and
`DELETE_TUPLE` undoflags[1].
- (unrelated) `WALWriteState::SwitchTable` no longer takes a rawpointer,
it was expected to be non-null anyways, it now takes a reference to
communicate this better.

[1] Important to note that `modified_tables` is still required to keep
the `shared_ptr<DataTable>` alive, so it can be referenced in the
cleanup state logic.
Since the DuckTableEntry that otherwise owns the `shared_ptr<DataTable>`
will have been deleted as part of the garbage collection before the
cleanup logic runs.

Upstream-commit: https://github.com/duckdb/duckdb/commit/07effc2ad0deca1fea8fd9abaf7fef679161a08f