Partially remove BoundQueryNode - have Bind directly emit a BoundStatement (https://redirect.github.com/duckdb/duckdb/pull/19289)
In the long-long ago we used to have two separate stages for binding and
planning - `Bind` would turn a `SQLStatement` into a
`BoundSQLStatement`, and `Plan` would take a `BoundSQLStatement` and
create a `LogicalOperator`.

At some point - we got rid of the `BoundSQLStatement` - but many of the
components from the old days are still there, such as `BoundQueryNode`.
This PR takes one step towards removing those components and having the
binding stage be a full single stage that directly emits a
`LogicalOperator` by having `Bind(QueryNode)` directly emit a
`BoundStatement` which contains a `LogicalOperator`. We move subqueries
from having a `BoundQueryNode` to having a `LogicalOperator`.

The `BoundQueryNode` itself is not fully removed (yet!) as it is still
required in some parts. In particular, for set operations and CTEs there
are still some cases where we need to interact with these. This is an
intermediate step in a full plan of getting everything to emit
LogicalOperators, which would simplify / clean up the binder
significantly and enable easier future reworks.

Upstream-commit: https://github.com/duckdb/duckdb/commit/f793ea27c6ca1ee2f1ca22aff86ba6586575d085