test: add test for CREATE TABLE with `my_schema.my_enum[]` (https://redirect.github.com/duckdb/duckdb/pull/19296)
Adds a test for https://github.com/duckdb/duckdb/issues/17041.

This test looks like it fails on 1.4 when run on shell.duckdb.org, but
currently on main it appears to pass.

I still need to try cherry-picking this test back to 1.4 in my local
machine to confirm that it does fail (which would confirm that the fix
happened sometime between 1.4 and main)

Upstream-commit: https://github.com/duckdb/duckdb/commit/3dc630e8e442dea8b11be34bec881ad534bc1a5f