Binder: BoundQueryNode cleanup (https://redirect.github.com/duckdb/duckdb/pull/19322)
Follow-up from https://github.com/duckdb/duckdb/pull/19293

We mostly remove BoundQueryNode fully, including removing BoundCTENode
and BoundRecursiveCTENode. For now BoundSelectNode and BoundSetOpNode
remain - but they are always explicitly returned (instead of having a
generic base class that is referenced).

There are two components that still depend on these:

* Set operations do some pretty complex name / binding resolution that
involves looking at expressions in child `SELECT` statements. This is
currently not feasible to do with the resulting `LogicalOperator`
structure
* Modifiers are shared across set / select statements and the binding
code for these is shared. This is much easier to pull apart but just
hasn't been done yet.

The former is the main blocker for fully removing all BoundQueryNode
code and fully moving to just returning a LogicalOperator for each of
the bind calls.

Upstream-commit: https://github.com/duckdb/duckdb/commit/17bf45bc2c75712bb345d8bb125ee04f7e36ac1e