[Parquet] Implement shredded `VARIANT` support for COPY TO parquet (https://redirect.github.com/duckdb/duckdb/pull/19219)
This PR continues where https://redirect.github.com/duckdb/duckdb/pull/19125 left off

Using the `SHREDDING` copy option it's possible to provide a structured
type for a VARIANT column in the projection list.
The `typed_value` of the VARIANT column in the Parquet file will
resemble this structured type and be populated if possible, falling back
to the binary-encoded `value` column when the type of the VARIANT value
doesn't match.

Example use:
```sql
COPY (
	FROM (VALUES
		({'a': 21::INTEGER, 'b':false::VARIANT),
		({'a': 42::INTEGER, 'd': 'test'}::VARIANT)
	)
) TO 'myfile.parquet' (
	shredding {
		a: 'STRUCT(a INTEGER, b VARIANT, c BOOLEAN)'
	}
)
```

### `SHREDDING` Copy Option

The following table lists the relation between the DuckDB type and the
VARIANT type
| LogicalTypeId  | VariantLogicalType                  |
|----------------|-------------------------------------|
| STRUCT         | OBJECT                              |
| LIST           | ARRAY                               |
| BOOLEAN        | BOOL_TRUE, BOOL_FALSE               |
| TINYINT        | INT8                                |
| SMALLINT       | INT16                               |
| INTEGER        | INT32                               |
| BIGINT         | INT64                               |
| FLOAT          | FLOAT                               |
| DOUBLE         | DOUBLE                              |
| DECIMAL        | DECIMAL                             |
| DATE           | DATE                                |
| TIME           | TIME_MICROS                         |
| TIMESTAMP_TZ   | TIMESTAMP_MICROS_TZ                 |
| TIMESTAMP      | TIMESTAMP_MICROS                    |
| TIMESTAMP_NS   | TIMESTAMP_NANOS                     |
| BLOB           | BLOB                                |
| VARCHAR        | VARCHAR                             |
| UUID           | UUID                                |
| VARIANT | - |

VARIANT is reserved for use in `STRUCT` / `LIST`.
When it's used inside of `STRUCT` it means that the OBJECT is shredded
on that field, but that field isn't shredded on a specific type.
Whn it's used inside of `LIST` it means that the Variant is shredded on
the ARRAY type, but the element isn't shredded on a specific type.

### Implementation Details

When a VARIANT column is detected, we check the copy options to see if a
shredded type is supplied for it, this will be used to inform the
`variant_to_parquet_variant` function call to transform to the parquet
variant shape.

`variant_to_parquet_variant` now takes an optional second parameter,
which should be of type VARCHAR, this is the stringified type (see the
explanation of the `SHREDDING` copy option above).

In `variant_to_parquet_variant` is where the shredding happens, using
the specified type we will determine whether a VARIANT value goes to
`value` or to the `typed_value`.
`value` is written row-by-row, while the populating of `typed_value` is
performed in a vectorized fashion.

### Misc
#### `FieldID`
I've split up the `FieldID` and `ChildFieldIDS` structs to
`parquet_field_id.hpp/cpp` out of the `parquet_extension.hpp/cpp` just
to separate it a little nicer.

#### `VariantColumnReader`
Fixed a bug in the `VariantColumnReader`, where the `is_field` wasn't
propagated far enough to allow both `value` and `typed_value` to be NULL
to indicate the field as being missing from the OBJECT.

#### `case_insensitive_string_map_t`
Added a `StringUtil::CIHash` overload for c-str.
Used by `case_insensitive_string_map_t` and
`case_insensitive_string_map_t`

#### `VariantUtils::FindChildValues`
Adjusted the function to take in a `ValidityMask` to mark which row(s)
are missing the child value instead of returning as soon as one of the
rows is missing the child value

### Future Work
Implement auto-detection of shredding types, currently this has to be
provided through the `SHREDDING` copy option, otherwise the VARIANT will
be written entirely unshredded.

Upstream-commit: https://github.com/duckdb/duckdb/commit/b83205f805568495733f8e3684fd8b851b2f0157