Binder: Rework the way set operations work so that we can use the standard binding flow (https://redirect.github.com/duckdb/duckdb/pull/19342)
Follow-up from https://github.com/duckdb/duckdb/pull/19322

This PR reworks the way that set operations bind so that they can use
the standard binding flow which takes a `BoundStatement` from the
children and returns a `BoundStatement` when finished. This does require
extending `BoundStatement` with several members, however:

```cpp
struct ExtraBoundInfo {
	SetOperationType setop_type = SetOperationType::NONE;
	vector<shared_ptr<Binder>> child_binders;
	vector<BoundStatement> bound_children;
	vector<unique_ptr<ParsedExpression>> original_expressions;
};
```

The reason this is required is because we have some quite involved logic
in how set operations are bound. In particular, when binding expressions
within the result modifiers (i.e. `ORDER BY` and such), we peer into the
select statements to figure out if there are any matches. This allows
queries like this to work:

```sql
SELECT i + 1 FROM (VALUES (42)) t(i) UNION ALL SELECT 84 ORDER BY i + 1;
```

As we can match the `i + 1` in the `ORDER BY` to the `i + 1` in the
select statement.

We also can't "just" look at the original parsed expressions in the
`SelectNode` - as they need to be qualified so e.g. statements like this
work:

```sql
SELECT i FROM (VALUES (42)) t(i) UNION ALL SELECT 84 ORDER BY t.i;
```

Instead of peering into the `BoundQueryNode` of the children - we now
opt to return the required information to bind set operations as part of
the `BoundStatement`, allowing the regular binding flow to be used.

Upstream-commit: https://github.com/duckdb/duckdb/commit/0e5bf48abf1cf37b5585306cafd5cf4fe271e618